refactor(Command): 책임을 `view`로 이동, 패키지 이동
